[refactor] Main.js
Main.js 내 weather 부분 -> component/specific/weatherInfo.js와 css로 분리
