+++ src/pages/Main.js
@@ -1,17 +1,12 @@
 /*Main.js*/
-import React, { useState, useEffect, useRef, useContext } from "react";
+import React, { useState, useEffect, useRef } from "react";
 import Header from "../components/common/Header/Header";
 import styles from "./Main.module.css";
-import { fetchForeCast, fetchForeCast2 } from "../pages/forecast";
 import { useNavigate } from "react-router-dom";
 import API from "../API/api";
 import SensorDetail from "../components/specific/sensorDetail/sensorDetail";
+import WeatherInfo from "../components/specific/weatherInfo/weatherInfo";
 /*images*/
-import sunIcon from "../assets/images/main/sun_icon.png";
-import cloudyIcon from "../assets/images/main/cloudy_icon.png";
-import semicloudyIcon from "../assets/images/main/semicloudy_icon.png";
-import cloudyrainIcon from "../assets/images/main/cloudyrain_icon.png";
-import weatherlocation from "../assets/images/main/location_icon.png";
 import logsensor from "../assets/images/main/sensor_icon.png";
 import currentbuilding from "../assets/images/main/currentBuilding_icon.png";
 import floorplan from "../assets/images/main/floorplan.png";
@@ -29,14 +24,14 @@ function Main() {
   const [fadeOut, setFadeOut] = useState(false);
   const [isFadingOut, setIsFadingOut] = useState(false);
   const [showMessage, setShowMessage] = useState(true);
-  const [dateTime, setDateTime] = useState(new Date());
-  const [forecast, setForecast] = useState(null);
-  const [forecast2, setForecast2] = useState(null);
-  const [loadingdata, setLoading] = useState(true);
-  const [hoveredFloor, setHoveredFloor] = useState(null); // 현재 호버 중인 층
-  const [allSensorData, setAllSensorData] = useState([]); // 모든 강의실 데이터 저장
-  const [nickname, setNickname] = useState("사용자"); // 닉네임 기본값 설정
-  const [sensorLogs, setSensorLogs] = useState([]); //이상수치값
+  const [dateTime] = useState(new Date());
+  const [forecast] = useState(null);
+  const [forecast2] = useState(null);
+  const [loadingdata] = useState(true);
+  const [hoveredFloor, setHoveredFloor] = useState(null);
+  const [allSensorData, setAllSensorData] = useState([]);
+  const [nickname, setNickname] = useState("사용자");
+  const [sensorLogs, setSensorLogs] = useState([]);
   const navigate = useNavigate();
 
   const getLevelColor = (level) => {
@@ -48,7 +43,7 @@ function Main() {
       case "YELLOW":
         return "#FFBD2E";
       default:
-        return "black"; // 기본 색상
+        return "black";
     }
   };
 
@@ -69,7 +64,6 @@ function Main() {
     }
   };
 
-  // 컴포넌트가 마운트될 때 데이터 가져오기
   useEffect(() => {
     fetchSensorLogs();
   }, []);
@@ -86,7 +80,6 @@ function Main() {
     setshowNoSensorPopup(false);
   };
 
-  // 센서 수치 이상 로그 렌더링 함수
   const renderSensorLogs = () => (
     <div
       style={{
@@ -113,7 +106,6 @@ function Main() {
             .toString()
             .padStart(2, "0")}`;
 
-          // 센서 타입에 따른 단위 설정
           const getUnit = (sensorType) => {
             switch (sensorType) {
               case "Temperature":
@@ -126,7 +118,7 @@ function Main() {
               case "AmbientNoise":
                 return "dB";
               default:
-                return ""; // 단위가 없는 경우
+                return "";
             }
           };
 
@@ -140,7 +132,7 @@ function Main() {
                 {log.building} {log.name}
                 <span
                   style={{
-                    color: getLevelColor(log.level), // level에 따른 색상 적용
+                    color: getLevelColor(log.level),
                     fontWeight: "bold",
                     marginLeft: "10px",
                   }}
@@ -183,7 +175,7 @@ function Main() {
   const handleBuildingClick = (building, buildingInfo) => {
     if (building == "신공학관") {
       setIsFadingOut(true);
-      setFadeOut(true); // 나머지 건물들을 fade out 시킴
+      setFadeOut(true);
       setTimeout(() => {
         setSelectedBuilding(building);
         setFadeOut(false);
@@ -197,70 +189,10 @@ function Main() {
     setPopupContent(null);
     setSelectedBuilding(null);
   };
-  // 날씨 안내 멘트 생성 함수
-  const getWeatherAdvice = () => {
-    if (!forecast || !forecast2) {
-      return "날씨 데이터를 불러오는 중입니다.";
-    }
-    if (forecast.rain === "1") {
-      return "오늘은 하루 종일 비가 예상됩니다\n외출 시 우산을 챙기세요";
-    } else if (forecast.rain === "3") {
-      return "눈이 오는 날입니다\n외출 시 따뜻하게 입으세요";
-    } else if (forecast.rain === "2") {
-      return "오늘은 비와 눈이 섞여 내릴 수 있습니다\n미끄럼에 주의하세요";
-    } else if (forecast.cloudy === "1") {
-      return "맑은 날씨입니다\n야외 활동을 즐기기 좋은 날이에요!";
-    } else if (forecast.cloudy === "3") {
-      return "햇살이 구름 사이로 드문드문 비추고 있어요\n산책이나 가벼운 야외 활동을 즐겨보세요!";
-    } else if (forecast.cloudy === "4") {
-      return "흐린 날씨입니다\n실내 활동을 계획해보세요";
-    } else if (forecast.humidity < 30) {
-      return "건조한 날씨입니다\n수분 보충에 유의하세요";
-    } else if (forecast2.maxTemp >= 30) {
-      return "무더운 날씨가 예상됩니다\n시원한 곳에서 휴식을 취하세요";
-    } else if (forecast2.minTemp <= 5) {
-      return "추운 날씨입니다\n따뜻한 옷차림을 준비하세요";
-    } else {
-      return "오늘의 날씨에 맞게 계획을 세워보세요!";
-    }
-  };
-  // Format the date and time
-  const formatDate = (date) => {
-    const days = ["일", "월", "화", "수", "목", "금", "토"];
-    const dayOfWeek = days[date.getDay()];
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-    const hours = String(date.getHours()).padStart(2, "0");
-    const minutes = String(date.getMinutes()).padStart(2, "0");
-    return `${year}년 ${month}월 ${day}일 ${dayOfWeek} KST ${hours}:${minutes}`;
-  };
-  // Fetch weather data
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const weatherData = await fetchForeCast(60, 127);
-        const weatherData2 = await fetchForeCast2(60, 127);
-        setForecast(weatherData);
-        setForecast2(weatherData2);
-        setLoading(false);
-      } catch (error) {
-        // console.error("Error fetching weather data: ", error);
-        setLoading(false);
-      }
-    };
-    fetchData();
-  }, []);
-  useEffect(() => {
-    const timer = setInterval(() => setDateTime(new Date()), 60000);
-    return () => clearInterval(timer);
-  }, []);
 
   const imageRef = useRef(null);
   const [hoveredIndex, setHoveredIndex] = useState(null);
-  const [sensorData, setSensorData] = useState(null); // 센서 데이터 상태 추가
-
-  // 좌표 저장
+  const [sensorData, setSensorData] = useState(null);
   const coordinates = [
     { building: "신공학관", id: "3115", x: 464, y: 442 },
     { building: "신공학관", id: "3173", x: 488, y: 482 },
@@ -272,7 +204,6 @@ function Main() {
   ];
 
   const handleClick = (id) => {
-    // 클릭 시 해당 강의실 페이지로 이동
     navigate(`/figures/${id}`);
   };
   const floorcoordinates = [
@@ -288,7 +219,6 @@ function Main() {
   useEffect(() => {
     const fetchAllSensorData = async () => {
       try {
-        // 모든 강의실 ID에 대한 데이터를 요청
         const responses = await Promise.all(
           coordinates.map(async (coord) => {
             const endpoint = `/api/sensorData/recent/classroom?building=${encodeURIComponent(
@@ -585,75 +515,12 @@ function Main() {
             className={styles.weatherAndLogs}
             style={{ position: "relative" }}
           >
-            <div className={styles.weatherInfo}>
-              <h3
-                style={{
-                  fontSize: "clamp(15px, 1.6vw, 20px)",
-                }}
-              >
-                <img
-                  src={weatherlocation}
-                  alt="현재 위치"
-                  className={styles.icons}
-                />
-                서울 중구 필동
-              </h3>
-
-              {loadingdata ? (
-                <p>Loading weather data...</p>
-              ) : forecast && forecast2 ? (
-                <>
-                  <div className={styles.weatherloading}>
-                    <img
-                      src={
-                        forecast.rain >= "1"
-                          ? cloudyrainIcon
-                          : forecast.cloudy === "1"
-                          ? sunIcon
-                          : forecast.cloudy === "3"
-                          ? semicloudyIcon
-                          : cloudyIcon
-                      }
-                      alt="날씨 이미지"
-                      className={styles.weatherImg}
-                    />
-                    <h1 className={styles.weatherTemperature}>
-                      {forecast.temperature}°C
-                    </h1>
-                  </div>
-                  <p className={styles.weatherRain}>
-                    {forecast.rain >= "1"
-                      ? "흐리고 비"
-                      : forecast.cloudy === "1"
-                      ? "맑음"
-                      : forecast.cloudy === "3"
-                      ? "구름많음"
-                      : "흐림"}
-                  </p>
-                  <p className={styles.weatherMinMax}>
-                    최고 {forecast2.maxTemp}°C / 최저 {forecast2.minTemp}°C
-                    <span style={{ marginLeft: "0.5vw" }}>
-                      습도 {forecast.humidity}%
-                    </span>
-                  </p>
-                  <p className={styles.weatherMinMax2}>
-                    {" "}
-                    {formatDate(dateTime)}
-                  </p>
-                  <p
-                    className={styles.weatherAdvice}
-                    style={{
-                      textAlign: "right",
-                    }}
-                  >
-                    <p className={styles.weatherAdvice}>{getWeatherAdvice()}</p>
-                  </p>
-                </>
-              ) : (
-                <p>데이터를 불러오는 중입니다.</p>
-              )}
-            </div>
-
+            <WeatherInfo
+              forecast={forecast}
+              forecast2={forecast2}
+              loading={loadingdata}
+              dateTime={dateTime}
+            />
             <div className={styles.sensorLogs}>
               <h3
                 style={{
@@ -696,7 +563,7 @@ function Main() {
       {showPopup && (
         <Popup
           popupContent="로그인이 필요한 서비스입니다."
-          onClose={closePopupHandler} // 팝업 닫기 함수
+          onClose={closePopupHandler}
           registerLink="/login"
           buttonText="로그인"
         />
@@ -706,15 +573,15 @@ function Main() {
         (token ? (
           <Popup
             popupContent="해당 건물에 등록된 센서가 없습니다.등록하시겠습니까?"
-            onClose={closeshowNoSensorPopupHandler} // 팝업 닫기 함수
-            registerLink="/manager" // 센서 등록 페이지로 링크
+            onClose={closeshowNoSensorPopupHandler}
+            registerLink="/manager"
             buttonText="등록하기"
           />
         ) : (
           <Popup
             popupContent="해당 건물에 등록된 센서가 없습니다.등록을 원하시면 로그인해주세요."
-            onClose={closeshowNoSensorPopupHandler} // 팝업 닫기 함수
-            registerLink="/login" // 홈 페이지로 링크
+            onClose={closeshowNoSensorPopupHandler}
+            registerLink="/login"
             buttonText="로그인"
           />
         ))}
